1. OptionMenu를 추가하여 새로고침 하는 기능 추가 2. 아래로 스와이프하여 새로고침 기능 추가

+++ app/src/main/java/com/leethcher/weatherapp/models/Sys.kt
@@ -8,6 +8,6 @@ data class Sys(
     val type: Int,
     val message: Double,
     val country: String,
-    val sunrise: Int,
-    val sunset: Int
+    val sunrise: Long,
+    val sunset: Long
 ): Parcelable
